Completely revamped translation tool
- Removed everything to do with comments. Next
version will have a discussion board for each
language. This is because discussions often are
more general around the language than specific to
certain phrases. Example of such discussion would
be around the level of politeness to use.

- Removed the approval phrase, and the translation
tool is more like a wiki now. However it's locked
down for only team leaders and administrators.