Initial work on new menu system
The problem with the previous menu system was
that it was:

1. Confusing to add new menu items. Ordinary menu
items, playlists, external user subscription all
had their own way of doing menu items.

2. It was coupled with the tab system which was
confusing and not necessary in any way.

3. The Menu "class" kept all the references to
menu items which led to hacky selector thingy
solutions when a model needed to update its menu
view.

4. Had a weird code design with a huge switch/case
instead of something more object oriented.

This commit targets these issues by making the
menu more of a part of the infrastructure. Now
other modules such as playlists and the home
screen can add menu items and keep references to
them so that they can be updated when the model
changes.

Also, the menu keeps track of selected menu items
so that there's no need for potentially slow
selectors in order to add/remove class names.